requirements.txt
discord
ffmpeg
python-dotenv
pillow